ENH: Enabled trimming of function and gradient in Linesearch
If the target function has singularities, it usually happens that the
linesearch hits one of them. The step will be rejected, but the high
function value is still used in polynomial_fit, resulting in bad step
sizes.

Also cleaned up NLCG stopping criteria.

+++ tools/dbg_nlopt/dbg_nlopt.cpp
@@ -253,7 +253,7 @@ int main(int argc, char* argv[])
   Runtime::initialise(argc, argv);
 
   typedef Mem::Main MemType;
-  typedef double DataType;
+  typedef float DataType;
   typedef unsigned int IndexType;
 
   // The analytic function we want to minimise. Look at the Analytic::Common namespace for other candidates.
